[SofaImGui] Refactor Windows menu bar methods (#109)
* [SofaImGui] Refactor methods

* [SofaImGui] Refactor methods

* [SofaImGui] Refactor method Display Flags

* [SofaImGui] Refactor method Plugins

* [SofaImGui] Refactor method Components

* [SofaImGui] Refactor method Settings

* [SofaImGui] Refactor method ViewPort

* [SofaImGui] Clean Unused Includes

* [SofaImGui] Removing useless Library Calls + Adding Doxygen comment

* [SofaImGui] Adding Doxygen comments

* [SofaImGui] Adding Doxygen comments

* [SofaImGui] Update Doxygen comments Positions

* [SofaImGui] Remove Useless blank lines
