+++ SofaImGui/CMakeLists.txt
@@ -84,6 +84,19 @@ set(HEADER_FILES
     ${SOFAIMGUI_SOURCE_DIR}/ImGuiGUI.h
     ${SOFAIMGUI_SOURCE_DIR}/ImGuiGUIEngine.h
     ${SOFAIMGUI_SOURCE_DIR}/ObjectColor.h
+    ${SOFAIMGUI_SOURCE_DIR}/windows/Performances.h
+    ${SOFAIMGUI_SOURCE_DIR}/windows/Log.h
+    ${SOFAIMGUI_SOURCE_DIR}/windows/Profiler.h
+    ${SOFAIMGUI_SOURCE_DIR}/windows/SceneGraph.h
+    ${SOFAIMGUI_SOURCE_DIR}/windows/DisplayFlags.h
+    ${SOFAIMGUI_SOURCE_DIR}/windows/Plugins.h
+    ${SOFAIMGUI_SOURCE_DIR}/windows/Components.h
+    ${SOFAIMGUI_SOURCE_DIR}/windows/Settings.h
+    ${SOFAIMGUI_SOURCE_DIR}/windows/ViewPort.h
+    ${SOFAIMGUI_SOURCE_DIR}/AppIniFile.h
+
+
+
 )
 
 set(SOURCE_FILES
@@ -92,6 +105,20 @@ set(SOURCE_FILES
     ${SOFAIMGUI_SOURCE_DIR}/ImGuiGUIEngine.cpp
     ${SOFAIMGUI_SOURCE_DIR}/ObjectColor.cpp
     ${SOFAIMGUI_SOURCE_DIR}/initSofaImGui.cpp
+    ${SOFAIMGUI_SOURCE_DIR}/windows/Performances.cpp
+    ${SOFAIMGUI_SOURCE_DIR}/windows/Log.cpp
+    ${SOFAIMGUI_SOURCE_DIR}/windows/Profiler.cpp
+    ${SOFAIMGUI_SOURCE_DIR}/windows/SceneGraph.cpp
+    ${SOFAIMGUI_SOURCE_DIR}/windows/DisplayFlags.cpp
+    ${SOFAIMGUI_SOURCE_DIR}/windows/Plugins.cpp
+    ${SOFAIMGUI_SOURCE_DIR}/windows/Components.cpp
+    ${SOFAIMGUI_SOURCE_DIR}/windows/Settings.cpp
+    ${SOFAIMGUI_SOURCE_DIR}/windows/ViewPort.cpp
+    ${SOFAIMGUI_SOURCE_DIR}/AppIniFile.cpp
+
+
+
+
 )
