+++ SofaImGui/src/SofaImGui/windows/Log.h
@@ -20,6 +20,8 @@
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
 #pragma once
+
+#include <sofa/simulation/Node.h>
 #include <SofaImGui/config.h>
 
 #include <memory>
@@ -30,38 +32,19 @@
 #include <sofa/simulation/Node.h>
 #include <SimpleIni.h>
 
-struct GLFWwindow;
-namespace sofa::glfw
+
+namespace windows
 {
-    class SofaGLFWBaseGUI;
-}
+        /**
+         * @brief Shows the Log window.
+         *
+         * This function displays a window containing log messages. It provides options to filter messages by type and save the log to a file. The displayed log messages include their IDs, types, senders, and the messages themselves.
+         *
+         * @param windowNameLog The name of the Log window.
+         * @param isLogWindowOpen A reference to a boolean flag indicating if the Log window is open.
+         */
+        void showLog(const char* const& windowNameLog,
+                     bool& isLogWindowOpen);
 
-namespace sofaimgui
-{
-
-class ImGuiGUIEngine : public sofaglfw::BaseGUIEngine
-{
-public:
-    ImGuiGUIEngine() = default;
-    ~ImGuiGUIEngine() = default;
-    
-    void init() override;
-    void initBackend(GLFWwindow*) override;
-    void startFrame(sofaglfw::SofaGLFWBaseGUI*) override;
-    void endFrame() override {}
-    void beforeDraw(GLFWwindow* window) override;
-    void afterDraw() override;
-    void terminate() override;
-    bool dispatchMouseEvents() override;
-
-protected:
-    std::unique_ptr<sofa::gl::FrameBufferObject> m_fbo;
-    std::pair<unsigned int, unsigned int> m_currentFBOSize;
-    std::pair<float, float> m_viewportWindowSize;
-    bool isMouseOnViewport { false };
-    CSimpleIniA ini;
-
-    void loadFile(sofaglfw::SofaGLFWBaseGUI* baseGUI, sofa::core::sptr<sofa::simulation::Node>& groot, std::string filePathName);
-};
 
 } // namespace sofaimgui
