+++ SofaImGui/src/SofaImGui/windows/Settings.h
@@ -20,48 +20,24 @@
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
 #pragma once
-#include <SofaImGui/config.h>
 
-#include <memory>
-#include <SofaGLFW/BaseGUIEngine.h>
-#include <sofa/gl/FrameBufferObject.h>
-
-#include <imgui.h>
 #include <sofa/simulation/Node.h>
-#include <SimpleIni.h>
 
-struct GLFWwindow;
-namespace sofa::glfw
-{
-    class SofaGLFWBaseGUI;
-}
 
-namespace sofaimgui
+namespace windows
 {
 
-class ImGuiGUIEngine : public sofaglfw::BaseGUIEngine
-{
-public:
-    ImGuiGUIEngine() = default;
-    ~ImGuiGUIEngine() = default;
-    
-    void init() override;
-    void initBackend(GLFWwindow*) override;
-    void startFrame(sofaglfw::SofaGLFWBaseGUI*) override;
-    void endFrame() override {}
-    void beforeDraw(GLFWwindow* window) override;
-    void afterDraw() override;
-    void terminate() override;
-    bool dispatchMouseEvents() override;
-
-protected:
-    std::unique_ptr<sofa::gl::FrameBufferObject> m_fbo;
-    std::pair<unsigned int, unsigned int> m_currentFBOSize;
-    std::pair<float, float> m_viewportWindowSize;
-    bool isMouseOnViewport { false };
-    CSimpleIniA ini;
-
-    void loadFile(sofaglfw::SofaGLFWBaseGUI* baseGUI, sofa::core::sptr<sofa::simulation::Node>& groot, std::string filePathName);
-};
+        /**
+         * @brief Shows the Settings window.
+         *
+         * This function displays settings for configuring the application, such as theme selection, global scale, and viewport settings.
+         *
+         * @param windowNameSettings The name of the Settings window.
+         * @param isSettingsOpen A reference to a boolean flag indicating if the Settings window is open.
+         * @param ini The INI file object containing application settings.
+         */
+        void showSettings(const char* const& windowNameSettings,
+                          bool& isSettingsOpen,
+                          CSimpleIniA &ini);
 
 } // namespace sofaimgui
