+++ SofaImGui/src/SofaImGui/windows/ViewPort.h
@@ -20,48 +20,33 @@
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
 #pragma once
-#include <SofaImGui/config.h>
 
-#include <memory>
-#include <SofaGLFW/BaseGUIEngine.h>
-#include <sofa/gl/FrameBufferObject.h>
-
-#include <imgui.h>
 #include <sofa/simulation/Node.h>
-#include <SimpleIni.h>
 
-struct GLFWwindow;
-namespace sofa::glfw
-{
-    class SofaGLFWBaseGUI;
-}
-
-namespace sofaimgui
+namespace windows
 {
 
-class ImGuiGUIEngine : public sofaglfw::BaseGUIEngine
-{
-public:
-    ImGuiGUIEngine() = default;
-    ~ImGuiGUIEngine() = default;
-    
-    void init() override;
-    void initBackend(GLFWwindow*) override;
-    void startFrame(sofaglfw::SofaGLFWBaseGUI*) override;
-    void endFrame() override {}
-    void beforeDraw(GLFWwindow* window) override;
-    void afterDraw() override;
-    void terminate() override;
-    bool dispatchMouseEvents() override;
-
-protected:
-    std::unique_ptr<sofa::gl::FrameBufferObject> m_fbo;
-    std::pair<unsigned int, unsigned int> m_currentFBOSize;
-    std::pair<float, float> m_viewportWindowSize;
-    bool isMouseOnViewport { false };
-    CSimpleIniA ini;
-
-    void loadFile(sofaglfw::SofaGLFWBaseGUI* baseGUI, sofa::core::sptr<sofa::simulation::Node>& groot, std::string filePathName);
-};
+        /**
+         * @brief Displays the viewport window.
+         *
+         * This function renders the viewport window, showing the scene rendered into a frame buffer object (FBO).
+         * It also provides options to show/hide grid, axis, and frame within the viewport.
+         *
+         * @param groot The root node of the scene to be rendered.
+         * @param windowNameViewport The name of the viewport window.
+         * @param isViewportWindowOpen A reference to a boolean flag indicating if the viewport window is open.
+         * @param ini The INI file object containing application settings.
+         * @param m_fbo The frame buffer object (FBO) used for rendering the scene.
+         * @param m_viewportWindowSize A reference to a pair representing the width and height of the viewport window.
+         * @param isMouseOnViewport A reference to a boolean flag indicating if the mouse cursor is over the viewport.
+         */
+         void showViewPort(sofa::core::sptr<sofa::simulation::Node> groot,
+                           const char* const& windowNameViewport,
+                           bool& isViewportWindowOpen,
+                           CSimpleIniA &ini,
+                           std::unique_ptr<sofa::gl::FrameBufferObject>& m_fbo,
+                           std::pair<float,
+                           float>& m_viewportWindowSize,
+                           bool & isMouseOnViewport );
 
 } // namespace sofaimgui
